moved connected sources list to its own component. adding medical sources filter component for left menu with pipe for code => category name. filter component can exand and shrink wiht animations
TODO: clear checkbox list when query changes.
