+++ frontend/src/app/components/medical-sources-connected/medical-sources-connected.component.html
@@ -1,0 +1,57 @@
+<h2 class="az-content-title">Connected Sources</h2>
+
+<div *ngIf="!loading else isLoadingTemplate" class="row">
+  <div *ngFor="let sourceInfo of connectedSourceList" class="col-sm-3 mg-b-20 px-3">
+    <div class="card h-100 d-flex align-items-center justify-content-center mt-3 mt-3 rounded-0 cursor-pointer">
+      <div (click)="openModal(contentModalRef, sourceInfo)" class="card-body">
+
+        <div class="h-100 d-flex align-items-center">
+          <img [src]="'assets/sources/'+(sourceInfo.metadata.brand_logo ? sourceInfo.metadata.brand_logo : sourceInfo.metadata.source_type+'.png')" [alt]="sourceInfo?.metadata.display" class="img-fluid">
+        </div>
+        <div *ngIf="status[sourceInfo.metadata?.source_type]" class="progress">
+          <div [style.width]="status[sourceInfo?.metadata?.source_type] == 'authorize' ? '33%' : '66%'" class="bg-indigo progress-bar progress-bar-striped progress-bar-animated" role="progressbar"></div>
+        </div>
+      </div>
+      <div class="card-footer text-center p-1" style="width:100%">
+        <small class="tx-gray-700">
+          {{sourceInfo?.metadata.display}}
+        </small>
+      </div>
+    </div>
+  </div>
+</div>
+
+
+<ng-template #contentModalRef let-modal>
+  <div class="modal-header">
+    <h4 class="modal-title" id="modal-basic-title">{{modalSelectedSourceListItem?.metadata["display"]}}</h4>
+    <button type="button" class="btn btn-close" aria-label="Close" (click)="modal.dismiss('Cross click')">
+      <span aria-hidden="true">×</span>
+    </button>
+  </div>
+
+  <div class="modal-body">
+    <h6>Manage Source</h6>
+    <p>Existing connections can be "Synced", "Reconnected" or "Deleted"</p>
+    <ul>
+      <li><p><strong>Sync</strong> - Download all resources from this healthcare provider, storing them securely in Fasten</p></li>
+      <li><p><strong>Reconnect</strong> - If your healthcare connection has expired, you can use this button to reconnect</p></li>
+      <li><p><strong>Delete</strong> - Delete all resources for this healthcare provider. This will ONLY effect data stored in Fasten</p></li>
+    </ul>
+  </div>
+
+  <div class="modal-footer">
+    <button (click)="sourceSyncHandler(modalSelectedSourceListItem.source)" type="button" class="btn btn-indigo">Sync</button>
+<!--    <button (click)="connectHandler($event, modalSelectedSourceListItem.source['source_type'])" type="button" class="btn btn-outline-light">Reconnect</button>-->
+    <button type="button" class="btn btn-outline-danger">Delete</button>
+    <button (click)="modal.dismiss('Close click')" type="button" class="btn btn-outline-light">Close</button>
+  </div>
+</ng-template>
+
+<ng-template #isLoadingTemplate>
+  <div class="row">
+    <div class="col-12">
+      <app-loading-spinner [loadingTitle]="'Please wait, loading sources...'"></app-loading-spinner>
+    </div>
+  </div>
+</ng-template>
