+++ frontend/src/app/components/medical-sources-filter/medical-sources-category-lookup.pipe.spec.ts
@@ -1,0 +1,8 @@
+import { MedicalSourcesCategoryLookupPipe } from './medical-sources-category-lookup.pipe';
+
+describe('MedicalSourcesCategoryLookupPipe', () => {
+  it('create an instance', () => {
+    const pipe = new MedicalSourcesCategoryLookupPipe();
+    expect(pipe).toBeTruthy();
+  });
+});
