+++ frontend/src/app/components/shared.module.ts
@@ -72,6 +72,8 @@ import { ReportMedicalHistoryExplanationOfBenefitComponent } from './report-medi
 import {GridstackComponent} from './gridstack/gridstack.component';
 import {GridstackItemComponent} from './gridstack/gridstack-item.component';
 import { MedicalSourcesFilterComponent } from './medical-sources-filter/medical-sources-filter.component';
+import { MedicalSourcesConnectedComponent } from './medical-sources-connected/medical-sources-connected.component';
+import { MedicalSourcesCategoryLookupPipe } from './medical-sources-filter/medical-sources-category-lookup.pipe';
 
 @NgModule({
   imports: [
@@ -156,72 +158,75 @@ import { MedicalSourcesFilterComponent } from './medical-sources-filter/medical-
     MediaComponent,
     ReportMedicalHistoryExplanationOfBenefitComponent,
     MedicalSourcesFilterComponent,
+    MedicalSourcesConnectedComponent,
+    MedicalSourcesCategoryLookupPipe,
   ],
-    exports: [
-      BinaryComponent,
-      ComponentsSidebarComponent,
-      DiagnosticReportComponent,
-      DocumentReferenceComponent,
-      FallbackComponent,
-      FhirResourceComponent,
-      FhirResourceOutletDirective,
-      ImmunizationComponent,
-      ListAdverseEventComponent,
-      ListAllergyIntoleranceComponent,
-      ListAppointmentComponent,
-      ListCarePlanComponent,
-      ListCommunicationComponent,
-      ListConditionComponent,
-      ListCoverageComponent,
-      ListDeviceComponent,
-      ListDeviceRequestComponent,
-      ListDiagnosticReportComponent,
-      ListDocumentReferenceComponent,
-      ListEncounterComponent,
-      ListGenericResourceComponent,
-      ListGoalComponent,
-      ListImmunizationComponent,
-      ListMedicationAdministrationComponent,
-      ListMedicationComponent,
-      ListMedicationDispenseComponent,
-      ListMedicationRequestComponent,
-      ListNutritionOrderComponent,
-      ListObservationComponent,
-      ListPatientComponent,
-      ListProcedureComponent,
-      ListServiceRequestComponent,
-      MedicalSourcesFilterComponent,
-      MedicationRequestComponent,
-      NlmTypeaheadComponent,
-      PractitionerComponent,
-      ProcedureComponent,
-      ReportHeaderComponent,
-      ReportLabsObservationComponent,
-      ReportMedicalHistoryConditionComponent,
-      ReportMedicalHistoryEditorComponent,
-      ReportMedicalHistoryExplanationOfBenefitComponent,
-      ResourceListComponent,
-      ResourceListOutletDirective,
-      ToastComponent,
-      UtilitiesSidebarComponent,
+  exports: [
+    BinaryComponent,
+    ComponentsSidebarComponent,
+    DiagnosticReportComponent,
+    DocumentReferenceComponent,
+    FallbackComponent,
+    FhirResourceComponent,
+    FhirResourceOutletDirective,
+    ImmunizationComponent,
+    ListAdverseEventComponent,
+    ListAllergyIntoleranceComponent,
+    ListAppointmentComponent,
+    ListCarePlanComponent,
+    ListCommunicationComponent,
+    ListConditionComponent,
+    ListCoverageComponent,
+    ListDeviceComponent,
+    ListDeviceRequestComponent,
+    ListDiagnosticReportComponent,
+    ListDocumentReferenceComponent,
+    ListEncounterComponent,
+    ListGenericResourceComponent,
+    ListGoalComponent,
+    ListImmunizationComponent,
+    ListMedicationAdministrationComponent,
+    ListMedicationComponent,
+    ListMedicationDispenseComponent,
+    ListMedicationRequestComponent,
+    ListNutritionOrderComponent,
+    ListObservationComponent,
+    ListPatientComponent,
+    ListProcedureComponent,
+    ListServiceRequestComponent,
+    MedicalSourcesFilterComponent,
+    MedicationRequestComponent,
+    NlmTypeaheadComponent,
+    PractitionerComponent,
+    ProcedureComponent,
+    ReportHeaderComponent,
+    ReportLabsObservationComponent,
+    ReportMedicalHistoryConditionComponent,
+    ReportMedicalHistoryEditorComponent,
+    ReportMedicalHistoryExplanationOfBenefitComponent,
+    ResourceListComponent,
+    ResourceListOutletDirective,
+    ToastComponent,
+    UtilitiesSidebarComponent,
 
-      //standalone components
-      BadgeComponent,
-      TableComponent,
-      CodingComponent,
-      LoadingSpinnerComponent,
-      GlossaryLookupComponent,
-      AllergyIntoleranceComponent,
-      MedicationComponent,
-      MedicationRequestComponent,
-      PractitionerComponent,
-      ProcedureComponent,
-      ImmunizationComponent,
-      BinaryComponent,
-      GridstackComponent,
-      GridstackItemComponent,
+    //standalone components
+    BadgeComponent,
+    TableComponent,
+    CodingComponent,
+    LoadingSpinnerComponent,
+    GlossaryLookupComponent,
+    AllergyIntoleranceComponent,
+    MedicationComponent,
+    MedicationRequestComponent,
+    PractitionerComponent,
+    ProcedureComponent,
+    ImmunizationComponent,
+    BinaryComponent,
+    GridstackComponent,
+    GridstackItemComponent,
+    MedicalSourcesConnectedComponent,
 
-    ]
+  ]
 })
 
 export class SharedModule { }
