+++ frontend/src/app/models/lighthouse/lighthouse-source-search.ts
@@ -1,5 +1,22 @@
 import {MetadataSource} from '../fasten/metadata-source';
 
+export class LighthouseSourceSearchResult {
+  _index: string;
+  _type: string;
+  _id: string;
+  _score: number;
+  _source: MetadataSource;
+  sort: string[];
+}
+
+export class LighthouseSourceSearchAggregation {
+  sum_other_doc_count: number;
+  buckets: {
+    key: string;
+    doc_count: number;
+  }[]
+}
+
 export class LighthouseSourceSearch {
   _scroll_id: string;
   took: number;
@@ -10,28 +27,10 @@ export class LighthouseSourceSearch {
       relation: string;
     };
     max_score: number;
-    hits: {
-      _index: string;
-      _type: string;
-      _id: string;
-      _score: number;
-      _source: MetadataSource;
-    }[];
+    hits: LighthouseSourceSearchResult[];
   };
   aggregations: {
-    by_platform_type: {
-      sum_other_doc_count: number;
-      buckets: {
-        key: string;
-        doc_count: number;
-      }[]
-    },
-    by_category: {
-      sum_other_doc_count: number;
-      buckets: {
-        key: string;
-        doc_count: number;
-      }[]
-    }
-  }
+    by_platform_type: LighthouseSourceSearchAggregation
+    by_category: LighthouseSourceSearchAggregation
+  };
 }
